+++ nevergrad/optimization/test_optimizerlib.py
@@ -55,7 +55,7 @@ def check_optimizer(
 ) -> None:
     # recast optimizer do not support num_workers > 1, and respect no_parallelization.
     num_workers = 1 if optimizer_cls.recast or optimizer_cls.no_parallelization else 2
-    num_attempts = 1 if not verify_value else 2  # allow 2 attemps to get to the optimum (shit happens...)
+    num_attempts = 1 if not verify_value else 3  # allow 3 attemps to get to the optimum (shit happens...)
     optimum = [0.5, -0.8]
     fitness = Fitness(optimum)
     for k in range(1, num_attempts + 1):
@@ -112,6 +112,7 @@ SLOW = [
     "ASCMA2PDEthird",
     "MultiScaleCMA",
     "PCEDA",
+    "EDA",
     "MicroCMA",
     "ES",
 ]
@@ -124,7 +125,7 @@ def test_optimizers(name: str) -> None:
     if isinstance(optimizer_cls, base.ConfiguredOptimizer):
         assert any(hasattr(mod, name) for mod in (optlib, xpvariants))  # make sure registration matches name in optlib/xpvariants
         assert optimizer_cls.__class__(**optimizer_cls._config) == optimizer_cls, "Similar configuration are not equal"
-    verify = not optimizer_cls.one_shot and name not in SLOW and not any(x in name for x in ["BO", "Discrete", "MutDE"])
+    verify = not optimizer_cls.one_shot and name not in SLOW and not any(x in name for x in ["BO", "Discrete"])
     # the following context manager speeds up BO tests
     patched = partial(acq_max, n_warmup=10000, n_iter=2)
     with patch("bayes_opt.bayesian_optimization.acq_max", patched):
@@ -164,6 +165,9 @@ def test_optimizers_suggest(name: str) -> None:  # pylint: disable=redefined-out
         candidate = optimizer.ask()
         try:
             optimizer.tell(candidate, 12)
+            # The optimizer should recommend its suggestion, except for a few optimization methods:
+            if name not in ["SPSA", "TBPSA", "StupidRandom"]:
+                np.testing.assert_array_almost_equal(optimizer.provide_recommendation().value, [12.0] * 4)
         except base.TellNotAskedNotSupportedError:
             pass
 
@@ -197,19 +201,22 @@ def test_optimizers_recommendation(name: str, recomkeeper: RecommendationKeeper)
         # Reducing the precision could help in this regard.
         patched = partial(acq_max, n_warmup=10000, n_iter=2)
         with patch("bayes_opt.bayesian_optimization.acq_max", patched):
-            candidate = optim.minimize(fitness)
+            recom = optim.minimize(fitness)
     if name not in recomkeeper.recommendations.index:
-        recomkeeper.recommendations.loc[name, :dimension] = tuple(candidate.args[0])
+        recomkeeper.recommendations.loc[name, :dimension] = tuple(recom.value)
         raise ValueError(f'Recorded the value for optimizer "{name}", please rerun this test locally.')
     # BO slightly differs from a computer to another
     decimal = 2 if isinstance(optimizer_cls, optlib.ParametrizedBO) or "BO" in name else 5
     np.testing.assert_array_almost_equal(
-        candidate.args[0],
+        recom.value,
         recomkeeper.recommendations.loc[name, :][:dimension],
         decimal=decimal,
         err_msg="Something has changed, if this is normal, delete the following "
         f"file and rerun to update the values:\n{recomkeeper.filepath}",
     )
+    # check that by default the recommendation has been evaluated
+    if isinstance(optimizer_cls, optlib.EvolutionStrategy):  # no noisy variants
+        assert recom.loss is not None
 
 
 @testing.parametrized(
@@ -357,6 +364,11 @@ def test_parametrization_optimizer_reproducibility() -> None:
     parametrization.random_state.seed(12)
     optimizer = optlib.RandomSearch(parametrization, budget=10)
     recom = optimizer.minimize(_square)
+    np.testing.assert_equal(recom.kwargs["y"], 4)
+    # resampling deterministically
+    # (this test has been reeeally useful so far, any change of the output must be investigated)
+    data = recom.get_standardized_data(reference=optimizer.parametrization)
+    recom = optimizer.parametrization.spawn_child().set_standardized_data(data, deterministic=True)
     np.testing.assert_equal(recom.kwargs["y"], 67)
 
 
@@ -379,7 +391,7 @@ def test_constrained_optimization() -> None:
     np.testing.assert_array_almost_equal([recom.kwargs["x"][0], recom.kwargs["y"]], [1.005573e+00, 3.965783e-04])
 
 
-@pytest.mark.parametrize("name", [name for name in registry])  # type: ignore
+@pytest.mark.parametrize("name", registry)  # type: ignore
 def test_parametrization_offset(name: str) -> None:
     if "PSO" in name or "BO" in name:
         raise SkipTest("PSO and BO have large initial variance")